backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
google-generativeai==0.4.1
faiss-cpu==1.7.4
numpy==1.26.1
openai-whisper
python-multipart==0.0.6
websockets==11.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
